requirements.txt
Babel==2.9.0
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.22.1
numpy==1.19.5
pandas==1.2.0
passlib==1.7.4
pymongo==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-http-client==3.3.1
pytz==2020.5
requests==2.25.1
sendgrid==6.5.0
six==1.15.0
speaklater==1.3
starkbank-ecdsa==1.1.0
urllib3==1.26.2
uWSGI==2.0.19.1
waitress==1.4.4
Werkzeug==1.0.1
WTForms==2.3.3
